moved KompasText to a separate file
